OBGM-47 Refactor UserExcelImporter - migrate extractDefaultRoles, createOrUpdateUser methods to userService
